+++ src/models/CID.ts
@@ -1,0 +1,4 @@
+import { OBSSStorage } from '@big-whale-labs/obss-storage-contract'
+
+type CID = OBSSStorage.CIDStructOutput
+export default CID
